Add support for Dreo dehumidifiers and integrate related functionality
