+++ custom_components/dreo/pydreo/models.py
@@ -112,7 +112,6 @@ SUPPORTED_MODEL_PREFIXES = {
     "WH",
     "DR-HAC",
     "DR-HHM",
-    "DR-HDH",
     "DR-HEC"
 }
 
@@ -295,7 +294,15 @@ SUPPORTED_DEVICES = {
     "DR-HHM": DreoDeviceDetails(device_type=DreoDeviceType.HUMIDIFIER),
 
     # Dehumidifiers
-    "DR-HDH001S": DreoDeviceDetails(
+        "DR-HDH001S": DreoDeviceDetails(
+        device_type=DreoDeviceType.DEHUMIDIFIER,
+        device_ranges={
+            HUMIDITY_RANGE: (30, 85),
+            SPEED_RANGE: (1, 3)
+        }
+    ),
+
+    "DR-HDH002S": DreoDeviceDetails(
         device_type=DreoDeviceType.DEHUMIDIFIER,
         device_ranges={
             HUMIDITY_RANGE: (30, 85),
