+++ tests/dreo/integrationtests/test_dreoaircirculator.py
@@ -82,7 +82,7 @@ class TestDreoAirCirculator(IntegrationTestBase):
             pydreo_fan = self.pydreo_manager.devices[0]
             ha_fan = fan.DreoFanHA(pydreo_fan)
             assert ha_fan.is_on is False
-            assert ha_fan.speed_count == 9
+            assert ha_fan.speed_count == 10
             assert not(ha_fan.supported_features & FanEntityFeature.PRESET_MODE)
 
             """This shouldn't be needed, but for some reason HA calls this even if preset_mode is not supported."""
